requirements.txt
affine==2.4.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.1
Fiona==1.9.2
fonttools==4.39.0
geopandas==0.12.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
keyboard==0.13.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
munch==2.5.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
rasterio==1.3.6
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.10.1
segment-anything==1.0
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
sympy==1.11.1
torch==2.0.0+cpu
torchaudio==2.0.0+cpu
torchvision==0.15.0+cpu
typing-extensions==4.5.0
urllib3==1.26.15
vext==0.7.6
vext.pyqt5==0.7.4
zipp==3.15.0
